Started the high-level functions #4 - Renamed YoutubeManagerV3 to YoutubeApiV3 -  child class YoutubeManager that used the mysql class and contains the high-level functions - Created the add_channel and remove_channel functions and integrated the current functionality into the main.py
