+++ youbot/__init__.py
@@ -5,7 +5,7 @@ from youbot.configuration import Configuration, validate_json_schema
 from youbot.cloudstore import DropboxCloudstore
 from youbot.datastore import YoutubeMySqlDatastore
 from youbot.emailer import GmailEmailer
-from youbot.youtube_utils import YoutubeManagerV3
+from youbot.youtube_utils import YoutubeManager, YoutubeApiV3
 
 __author__ = "drkostas"
 __email__ = "georgiou.kostas94@gmail.com"
